Fix test.
JAVA-5650

+++ driver-core/src/test/functional/com/mongodb/client/model/search/AggregatesVectorSearchIntegrationTest.java
@@ -179,7 +179,9 @@ class AggregatesVectorSearchIntegrationTest {
 
     @AfterAll
     static void afterAll() {
-        collectionHelper.drop();
+        if (collectionHelper != null) {
+            collectionHelper.drop();
+        }
     }
 
     private static Stream<Arguments> provideSupportedVectors() {
